prompt.jsをactivityPrompt.jsに変更し、selectの質問をintensitySelector.jsに分離
